fix features
